use AsSingleRow everywhere & small fixes

+++ src/back/space/src/repo/table_space.hpp
@@ -24,7 +24,7 @@ public:
 	void InsertDataForMocks();
 	bool IsExists(const std::string& key);
 	bool IsReadyForCreationByTime(const std::string& userId);
-	int GetCountSpacesWithUser(const std::string& userId);
+	int64_t GetCountSpacesWithUser(const std::string& userId);
 	void Delete(const boost::uuids::uuid& spaceUuid);
 	model::Space SelectById(const boost::uuids::uuid& id);
 	model::Space SelectByKey(const std::string& key);
